requirements.txt
requests
pandas
python-dotenv
matplotlib
scikit-learn
jupyter
streamlit
